Perf tests refactor to separate files, added initial app tests.
