+++ frontend/cypress/fixtures/perf/appsParams.json
@@ -1,0 +1,5 @@
+{
+  "namespaces": "bookinfo,alpha,beta,sleep",
+  "duration": "300",
+  "refresh": "15000"
+}
