+++ frontend/cypress/perf/constants.ts
@@ -1,0 +1,1 @@
+export const reportFilePath = 'logs/performance.txt';
